+++ piecrust-uplink/src/debug.rs
@@ -4,80 +4,43 @@
 //
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
-use core::fmt;
-use core::fmt::Write;
-use core::ptr::slice_from_raw_parts_mut;
+use core::fmt::{self, Write};
 
-pub mod ext {
-    extern "C" {
-        pub fn host_debug(ofs: i32, len: u32);
-    }
+use crate::{state, ARGBUF_LEN};
+
+extern "C" {
+    pub fn hdebug(arg_len: u32);
 }
 
-const DEBUG_BUFFER_SIZE: usize = 16 * 1024;
-static mut DEBUG_BUFFER: [u8; DEBUG_BUFFER_SIZE] = [0u8; DEBUG_BUFFER_SIZE];
+/// A small struct that can `fmt::Write` to the argument buffer.
+///
+/// It is just an offset to the argument buffer, representing how much has been
+/// written to it.
+#[derive(Default)]
+pub struct ArgbufWriter(usize);
 
-/// Write a string to the debug buffer and report it to the host.
-pub fn debug(s: &str) -> fmt::Result {
-    let mut w = DebugWriter::new();
-    w.write_str(s)?;
-    unsafe { ext::host_debug(w.ptr as i32, w.ofs as u32) };
-    Ok(())
-}
-
-/// A small struct that can `fmt::Write` to the debug buffer
-#[derive(Debug)]
-pub struct DebugWriter {
-    ptr: *mut u8,
-    ofs: usize,
-}
-
-impl DebugWriter {
-    /// Creates a new `DebugWriter`
-    pub fn new() -> Self {
-        DebugWriter {
-            ptr: unsafe { DEBUG_BUFFER.as_mut_ptr() },
-            ofs: 0,
-        }
-    }
-
-    pub fn ptr(&self) -> *mut u8 {
-        self.ptr
-    }
-
+impl ArgbufWriter {
     pub fn ofs(&self) -> usize {
-        self.ofs
-    }
-
-    fn slice(&mut self) -> Option<&mut [u8]> {
-        // if there is no space left, return `None`
-        if self.ofs < DEBUG_BUFFER_SIZE {
-            unsafe {
-                Some(&mut *slice_from_raw_parts_mut(
-                    self.ptr.add(self.ofs),
-                    DEBUG_BUFFER_SIZE - self.ofs,
-                ))
-            }
-        } else {
-            None
-        }
+        self.0
     }
 }
 
-impl Default for DebugWriter {
-    fn default() -> Self {
-        DebugWriter::new()
-    }
-}
-
-impl fmt::Write for DebugWriter {
+impl Write for ArgbufWriter {
     fn write_str(&mut self, s: &str) -> fmt::Result {
         let bytes = s.as_bytes();
+        let bytes_len = bytes.len();
 
-        let slice = self.slice().ok_or(fmt::Error)?;
-        slice[..bytes.len()].copy_from_slice(bytes);
+        let new_ofs = self.0 + bytes_len;
 
-        self.ofs += bytes.len();
+        if new_ofs > ARGBUF_LEN {
+            return Err(fmt::Error);
+        }
+
+        state::with_arg_buf(|buf| {
+            buf[self.0..new_ofs].copy_from_slice(bytes);
+        });
+
+        self.0 = new_ofs;
 
         Ok(())
     }
@@ -90,11 +53,9 @@ macro_rules! debug {
         #[allow(unused)]
         use core::fmt::Write as _;
 
-        let mut w = $crate::debug::DebugWriter::new();
+        let mut w = $crate::debug::ArgbufWriter::default();
         write!(&mut w, $($tt)*).unwrap();
 
-        unsafe {
-            $crate::debug::ext::host_debug(w.ptr() as i32, w.ofs() as u32)
-        }
+        unsafe { $crate::debug::hdebug(w.ofs() as u32) };
     };
 }